// src/models/EntitiesModel.ts
export interface INewUser {
  firstName: string;
  middleName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  dateOfBirth: string;
  state: string;
  city: string;
  district: string;
  ward: string;
  address: string;
  clinicId: string;
  activationCode: string;
  dialCode: string;
  country: string;
}